requirements.txt
boto3~=1.34.114
python-dotenv~=1.0.1
